🚛 move stories to components dir
