+++ .storybook/config.js
@@ -12,4 +12,4 @@ addDecorator(withKnobs);
 addDecorator(withTemplate);
 addDecorator(withVuetify);
 
-configure(require.context("./stories", true, /\.stories\.js$/), module);
+configure(require.context("../src/components", true, /\.stories\.js$/), module);
